[Build] Fixed MSVC build issues, ~500 warnings down to 1.

+++ Source/Backends/Vulkan/Layer/Source/Export/ShaderExportStreamer.cpp
@@ -275,7 +275,7 @@ bool ShaderExportStreamer::ProcessSegment(ShaderExportStreamSegment *segment) {
         // Copy into stream
         MessageStream messageStream;
         messageStream.SetSchema(streamInfo.typeInfo.messageSchema);
-        messageStream.SetData(stream, size, size / streamInfo.typeInfo.typeSize);
+        messageStream.SetData(stream, size, static_cast<uint32_t>(size / streamInfo.typeInfo.typeSize));
 
         // Add output
         output->AddStream(messageStream);
